backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
autoflake==2.3.1
bcrypt==4.2.1
black==24.10.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
logger==1.4
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pydantic==2.10.3
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
